Another fixup for rB6cdb3845 (Added collection props getter/setter)
Values were getting clamped

+++ source/blender/blenkernel/BKE_layer.h
@@ -110,7 +110,7 @@ void BKE_collection_engine_property_add_float(struct CollectionEngineSettings *c
 void BKE_collection_engine_property_add_int(struct CollectionEngineSettings *ces, const char *name, int value);
 struct CollectionEngineProperty *BKE_collection_engine_property_get(struct CollectionEngineSettings *ces, const char *name);
 int BKE_collection_engine_property_value_get_int(struct CollectionEngineSettings *ces, const char *name);
-int BKE_collection_engine_property_value_get_float(struct CollectionEngineSettings *ces, const char *name);
+float BKE_collection_engine_property_value_get_float(struct CollectionEngineSettings *ces, const char *name);
 void BKE_collection_engine_property_value_set_int(struct CollectionEngineSettings *ces, const char *name, int value);
 void BKE_collection_engine_property_value_set_float(struct CollectionEngineSettings *ces, const char *name, float value);
 bool BKE_collection_engine_property_use_get(struct CollectionEngineSettings *ces, const char *name);
